Added a new class, empty for now
Static variables currently used aren't ideal, and are only going to
complicate the code further. However, they are necessary for resuming
the game when the application itself isn't finished. The solution
proposed with this commit is to add all the objects we want to save
between sessions as static members of this class, and then handle
everything related to it within that class.