A lot of changes regarding static variables
Lots of changes regarding static variables, created two new static
classes, SinglePlayerObjects and MultiPlayerObjects, that serve to
remember settings of a single, or multi, player game, even after that
activity was destroyed. Also redid some of the constructors and internal
game logic to be more in line with that approach.